requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
python-dotenv
mysql-connector-python>=8.1.0
flask-wtf==1.2.1
cryptography
email-validator
streamlit
pandas
